requirements.txt
pyspark>=3.2.0
pandas>=1.5.0
openpyxl>=3.0.0
scikit-learn>=1.0.0
apache-airflow>=2.7.0
mysql-connector-python>=8.0.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
